requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
diff-match-patch==20230430
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.6
django-environ==0.11.2
django-import-export==4.0.5
et-xmlfile==1.1.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
install-jdk==1.1.0
JPype1==1.5.0
lxml==5.2.2
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
panadas==0.2
pandas==2.2.2
pillow==10.3.0
plum-dispatch==1.7.4
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
tabula-py==2.9.3
tabulate==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.6
whitenoise==6.6.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
